add tls layer, fix websocket hostname
