+++ proxy/client/config.go
@@ -18,11 +18,16 @@ type ShadowsocksConfig struct {
 	Enabled bool `json,yaml:"enabled"`
 }
 
+type TransportPluginConfig struct {
+	Enabled bool `json,yaml:"enabled"`
+}
+
 type Config struct {
-	Mux         MuxConfig         `json,yaml:"mux"`
-	Websocket   WebsocketConfig   `json,yaml:"websocket"`
-	Router      RouterConfig      `json,yaml:"router"`
-	Shadowsocks ShadowsocksConfig `json,yaml:"shadowsocks"`
+	Mux             MuxConfig             `json,yaml:"mux"`
+	Websocket       WebsocketConfig       `json,yaml:"websocket"`
+	Router          RouterConfig          `json,yaml:"router"`
+	Shadowsocks     ShadowsocksConfig     `json,yaml:"shadowsocks"`
+	TransportPlugin TransportPluginConfig `json:"transport_plugin" yaml:"transport-plugin"`
 }
 
 func init() {
