+++ proxy/server/server.go
@@ -5,11 +5,12 @@ import (
 	"github.com/p4gefau1t/trojan-go/config"
 	"github.com/p4gefau1t/trojan-go/proxy"
 	"github.com/p4gefau1t/trojan-go/proxy/client"
+	"github.com/p4gefau1t/trojan-go/tunnel/freedom"
 	"github.com/p4gefau1t/trojan-go/tunnel/mux"
-	"github.com/p4gefau1t/trojan-go/tunnel/raw"
 	"github.com/p4gefau1t/trojan-go/tunnel/router"
 	"github.com/p4gefau1t/trojan-go/tunnel/shadowsocks"
 	"github.com/p4gefau1t/trojan-go/tunnel/simplesocks"
+	"github.com/p4gefau1t/trojan-go/tunnel/tls"
 	"github.com/p4gefau1t/trojan-go/tunnel/transport"
 	"github.com/p4gefau1t/trojan-go/tunnel/trojan"
 	"github.com/p4gefau1t/trojan-go/tunnel/websocket"
@@ -24,16 +25,21 @@ func init() {
 		if err != nil {
 			return nil, err
 		}
-		clientStack := []string{raw.Name}
+		clientStack := []string{freedom.Name}
 		if cfg.Router.Enabled {
-			clientStack = []string{raw.Name, router.Name}
+			clientStack = []string{freedom.Name, router.Name}
 		}
+
 		root := &proxy.Node{
 			Name:       transport.Name,
 			Next:       make(map[string]*proxy.Node),
 			IsEndpoint: false,
 			Context:    ctx,
 			Server:     s,
+		}
+
+		if !cfg.TransportPlugin.Enabled {
+			root = root.BuildNext(tls.Name)
 		}
 
 		trojanSubTree := root
