+++ tunnel/freedom/conn.go
@@ -1,4 +1,4 @@
-package raw
+package freedom
 
 import (
 	"net"
