+++ tunnel/simplesocks/simplesocks_test.go
@@ -3,24 +3,32 @@ package simplesocks
 import (
 	"context"
 	"fmt"
+	"github.com/p4gefau1t/trojan-go/config"
+	"github.com/p4gefau1t/trojan-go/tunnel/transport"
 	"testing"
 
 	"github.com/p4gefau1t/trojan-go/common"
 	"github.com/p4gefau1t/trojan-go/test/util"
 	"github.com/p4gefau1t/trojan-go/tunnel"
-	"github.com/p4gefau1t/trojan-go/tunnel/raw"
 )
 
 func TestSimpleSocks(t *testing.T) {
-	addr := tunnel.NewAddressFromHostPort("tcp", "127.0.0.1", common.PickPort("tcp", "127.0.0.1"))
-	rawClient := &raw.FixedClient{
-		FixedAddr: addr,
+	port := common.PickPort("tcp", "127.0.0.1")
+	transportConfig := &transport.Config{
+		LocalHost:  "127.0.0.1",
+		LocalPort:  port,
+		RemoteHost: "127.0.0.1",
+		RemotePort: port,
 	}
-	rawServer, err := raw.NewServer(addr)
+	ctx := config.WithConfig(context.Background(), transport.Name, transportConfig)
+	tcpClient, err := transport.NewClient(ctx, nil)
 	common.Must(err)
-	c, err := NewClient(context.Background(), rawClient)
+	tcpServer, err := transport.NewServer(ctx, nil)
 	common.Must(err)
-	s, err := NewServer(context.Background(), rawServer)
+
+	c, err := NewClient(ctx, tcpClient)
+	common.Must(err)
+	s, err := NewServer(ctx, tcpServer)
 	common.Must(err)
 
 	conn1, err := c.DialConn(&tunnel.Address{
@@ -29,14 +37,13 @@ func TestSimpleSocks(t *testing.T) {
 		Port:        443,
 	}, nil)
 	common.Must(err)
-	conn1.Write([]byte("12345678"))
+	defer conn1.Close()
+	conn1.Write(util.GeneratePayload(1024))
 	conn2, err := s.AcceptConn(nil)
 	common.Must(err)
-	buf := [8]byte{}
+	defer conn2.Close()
+	buf := [1024]byte{}
 	common.Must2(conn2.Read(buf[:]))
-	if string(buf[:]) != "12345678" {
-		t.Fail()
-	}
 	if !util.CheckConn(conn1, conn2) {
 		t.Fail()
 	}
@@ -49,7 +56,9 @@ func TestSimpleSocks(t *testing.T) {
 			Port:        443,
 		},
 	})
+	defer packet1.Close()
 	packet2, err := s.AcceptPacket(nil)
+	defer packet2.Close()
 	_, m, err := packet2.ReadWithMetadata(buf[:])
 	common.Must(err)
 	fmt.Println(m)
@@ -57,4 +66,6 @@ func TestSimpleSocks(t *testing.T) {
 	if !util.CheckPacketOverConn(packet1, packet2) {
 		t.Fail()
 	}
+	s.Close()
+	c.Close()
 }
