+++ tunnel/tls/client.go
@@ -1,0 +1,163 @@
+package tls
+
+import (
+	"context"
+	"crypto/tls"
+	"crypto/x509"
+	"encoding/pem"
+	"github.com/p4gefau1t/trojan-go/common"
+	"github.com/p4gefau1t/trojan-go/config"
+	"github.com/p4gefau1t/trojan-go/log"
+	"github.com/p4gefau1t/trojan-go/tunnel"
+	"github.com/p4gefau1t/trojan-go/tunnel/tls/fingerprint"
+	"github.com/p4gefau1t/trojan-go/tunnel/transport"
+	"github.com/p4gefau1t/trojan-go/tunnel/websocket"
+	utls "github.com/refraction-networking/utls"
+	"io"
+	"io/ioutil"
+	"strings"
+)
+
+// Client is a tls client
+type Client struct {
+	verify        bool
+	sni           string
+	ca            *x509.CertPool
+	cipher        []uint16
+	sessionTicket bool
+	reuseSession  bool
+	fingerprint   string
+	keyLogger     io.WriteCloser
+	underlay      tunnel.Client
+}
+
+func (c *Client) Close() error {
+	if c.keyLogger != nil {
+		c.keyLogger.Close()
+	}
+	return c.underlay.Close()
+}
+
+func (c *Client) DialPacket(tunnel.Tunnel) (tunnel.PacketConn, error) {
+	panic("not supported")
+}
+
+func (c *Client) DialConn(_ *tunnel.Address, overlay tunnel.Tunnel) (tunnel.Conn, error) {
+	conn, err := c.underlay.DialConn(nil, &Tunnel{})
+	if err != nil {
+		return nil, common.NewError("tls failed to dial conn").Base(err)
+	}
+
+	if c.fingerprint != "" {
+		// utls fingerprint
+		tlsConn := utls.UClient(conn, &utls.Config{
+			RootCAs:            c.ca,
+			ServerName:         c.sni,
+			InsecureSkipVerify: !c.verify,
+			KeyLogWriter:       c.keyLogger,
+		}, utls.HelloCustom)
+		_, isWebsocket := overlay.(*websocket.Tunnel)
+		// TODO fix websocket ALPN problem
+		spec, err := fingerprint.GetClientHelloSpec(c.fingerprint, isWebsocket)
+		if err != nil {
+			return nil, common.NewError("invalid hello spec").Base(err)
+		}
+		if err := tlsConn.ApplyPreset(spec); err != nil {
+			return nil, common.NewError("tls failed to apply preset fingerprint").Base(err)
+		}
+		if err := tlsConn.Handshake(); err != nil {
+			return nil, common.NewError("tls failed to handshake with remote server").Base(err)
+		}
+		return &transport.Conn{
+			Conn: tlsConn,
+		}, nil
+	}
+	// golang default tls library
+	tlsConn := tls.Client(conn, &tls.Config{
+		InsecureSkipVerify:     !c.verify,
+		ServerName:             c.sni,
+		RootCAs:                c.ca,
+		KeyLogWriter:           c.keyLogger,
+		CipherSuites:           c.cipher,
+		SessionTicketsDisabled: !c.sessionTicket,
+	})
+	err = tlsConn.Handshake()
+	if err != nil {
+		return nil, common.NewError("tls failed to handshake with remote server").Base(err)
+	}
+	return &transport.Conn{
+		Conn: tlsConn,
+	}, nil
+}
+
+// NewClient creates a tls client
+func NewClient(ctx context.Context, underlay tunnel.Client) (*Client, error) {
+	cfg := config.FromContext(ctx, Name).(*Config)
+
+	if cfg.TLS.Fingerprint != "" {
+		_, err := fingerprint.GetClientHelloSpec(cfg.TLS.Fingerprint, cfg.Websocket.Enabled)
+		if err != nil {
+			return nil, err
+		}
+		log.Info("tls fingerprint", cfg.TLS.Fingerprint, "applied")
+	}
+
+	if cfg.TLS.SNI == "" {
+		cfg.TLS.SNI = cfg.RemoteHost
+		log.Warn("tls sni is unspecified")
+	}
+
+	client := &Client{
+		underlay:      underlay,
+		verify:        cfg.TLS.Verify,
+		sni:           cfg.TLS.SNI,
+		cipher:        fingerprint.ParseCipher(strings.Split(cfg.TLS.Cipher, ":")),
+		sessionTicket: cfg.TLS.ReuseSession,
+		fingerprint:   cfg.TLS.Fingerprint,
+	}
+
+	if cfg.TLS.CertPath != "" {
+		caCertByte, err := ioutil.ReadFile(cfg.TLS.CertPath)
+		if err != nil {
+			return nil, common.NewError("failed to load cert file").Base(err)
+		}
+		client.ca = x509.NewCertPool()
+		ok := client.ca.AppendCertsFromPEM(caCertByte)
+		if !ok {
+			log.Warn("invalid cert list")
+		}
+		log.Info("using custom cert")
+
+		// print cert info
+		pemCerts := caCertByte
+		for len(pemCerts) > 0 {
+			var block *pem.Block
+			block, pemCerts = pem.Decode(pemCerts)
+			if block == nil {
+				break
+			}
+			if block.Type != "CERTIFICATE" || len(block.Headers) != 0 {
+				continue
+			}
+			cert, err := x509.ParseCertificate(block.Bytes)
+			if err != nil {
+				continue
+			}
+			log.Trace("issuer:", cert.Issuer, "subject:", cert.Subject)
+		}
+	} else if len(cfg.TLS.CertBytes) != 0 {
+		client.ca = x509.NewCertPool()
+		ok := client.ca.AppendCertsFromPEM(cfg.TLS.CertBytes)
+		if !ok {
+			log.Warn("invalid cert list")
+		}
+		log.Info("using custom cert (data)")
+	}
+
+	if cfg.TLS.CertPath == "" && len(cfg.TLS.CertBytes) == 0 {
+		log.Info("cert is unspecified, using default ca list")
+	}
+
+	log.Debug("tls client created")
+	return client, nil
+}
