+++ tunnel/tls/tunnel.go
@@ -1,0 +1,26 @@
+package tls
+
+import (
+	"context"
+	"github.com/p4gefau1t/trojan-go/tunnel"
+)
+
+const Name = "TLS"
+
+type Tunnel struct{}
+
+func (t *Tunnel) Name() string {
+	return Name
+}
+
+func (t *Tunnel) NewClient(ctx context.Context, client tunnel.Client) (tunnel.Client, error) {
+	return NewClient(ctx, client)
+}
+
+func (t *Tunnel) NewServer(ctx context.Context, server tunnel.Server) (tunnel.Server, error) {
+	return NewServer(ctx, server)
+}
+
+func init() {
+	tunnel.RegisterTunnel(Name, &Tunnel{})
+}
