+++ tunnel/trojan/trojan_test.go
@@ -1,16 +1,18 @@
 package trojan
 
 import (
+	"bytes"
 	"context"
 	"fmt"
-	"testing"
-
+	"github.com/p4gefau1t/trojan-go/common"
+	"github.com/p4gefau1t/trojan-go/config"
+	"github.com/p4gefau1t/trojan-go/redirector"
+	"github.com/p4gefau1t/trojan-go/statistic"
 	"github.com/p4gefau1t/trojan-go/test/util"
 	"github.com/p4gefau1t/trojan-go/tunnel"
-	"github.com/p4gefau1t/trojan-go/tunnel/raw"
-
-	"github.com/p4gefau1t/trojan-go/common"
-	"github.com/p4gefau1t/trojan-go/statistic"
+	"github.com/p4gefau1t/trojan-go/tunnel/transport"
+	"net"
+	"testing"
 )
 
 type MockUser struct {
@@ -46,20 +48,29 @@ func (*MockAuth) ListUsers() []statistic.User {
 }
 
 func TestTrojan(t *testing.T) {
-	addr := tunnel.NewAddressFromHostPort("tcp", "127.0.0.1", common.PickPort("tcp", "127.0.0.1"))
-	tcpClient := &raw.FixedClient{
-		FixedAddr: addr,
+	port := common.PickPort("tcp", "127.0.0.1")
+	transportConfig := &transport.Config{
+		LocalHost:  "127.0.0.1",
+		LocalPort:  port,
+		RemoteHost: "127.0.0.1",
+		RemotePort: port,
 	}
-	tcpServer, err := raw.NewServer(addr)
-	ctx := context.Background()
+	ctx := config.WithConfig(context.Background(), transport.Name, transportConfig)
+	tcpClient, err := transport.NewClient(ctx, nil)
+	common.Must(err)
+	tcpServer, err := transport.NewServer(ctx, nil)
+	common.Must(err)
+	ctx, cancel := context.WithCancel(ctx)
 	s := &Server{
 		underlay:   tcpServer,
 		auth:       &MockAuth{},
-		ctx:        ctx,
-		redirAddr:  nil,
+		redirAddr:  tunnel.NewAddressFromHostPort("tcp", "127.0.0.1", util.EchoPort),
 		connChan:   make(chan tunnel.Conn, 32),
 		muxChan:    make(chan tunnel.Conn, 32),
 		packetChan: make(chan tunnel.PacketConn, 32),
+		ctx:        ctx,
+		cancel:     cancel,
+		redir:      redirector.NewRedirector(ctx),
 	}
 	go s.acceptLoop()
 	c := &Client{
@@ -100,4 +111,23 @@ func TestTrojan(t *testing.T) {
 	if !util.CheckPacketOverConn(packet1, packet2) {
 		t.Fail()
 	}
+
+	conn1.Close()
+	conn2.Close()
+	packet1.Close()
+	packet2.Close()
+
+	//redirecting
+	conn, err := net.Dial("tcp", fmt.Sprintf("127.0.0.1:%d", port))
+	common.Must(err)
+	sendBuf := util.GeneratePayload(1024)
+	recvBuf := [1024]byte{}
+	common.Must2(conn.Write(sendBuf))
+	common.Must2(conn.Read(recvBuf[:]))
+	if bytes.Equal(sendBuf, recvBuf[:]) {
+		t.Fail()
+	}
+	conn.Close()
+	c.Close()
+	s.Close()
 }
